1. Set speed and send configuration functionalities added to serial 2. reorganisation of serial utilities (moved to separate files and namespaces)
